Vault provider
A regression was introduced with the caching layer for the locally stored file. The code is changing the initialization for FileWriterOnce to only use the settings directory as opposed to the folder + vault path.

Fix the unit tests
